note down TODO about `dyn ServiceFactory<_>` and move headers + service builders into new server module
