+++ src/bin/src/main.rs
@@ -1,15 +1,11 @@
 // Copyright 2025 Sebastian Dobe <sebastiandobe@mailbox.org>
 
 use crate::logging::setup_logging;
-use actix_web::{middleware, web};
+use actix_web::web;
 use cryptr::EncKeys;
 use rauthy_common::constants::{BUILD_TIME, RAUTHY_VERSION};
 use rauthy_common::{HTTP_CLIENT, is_sqlite, password_hasher};
 use rauthy_handlers::generic::I18N_CONFIG;
-use rauthy_handlers::{
-    api_keys, auth_providers, blacklist, clients, dev_only, events, fed_cm, generic, groups, html,
-    oidc, roles, scopes, sessions, themes, users,
-};
 use rauthy_models::app_state::AppState;
 use rauthy_models::database::{Cache, DB};
 use rauthy_models::email;
@@ -244,205 +240,4 @@ https://github.com/sebadob/rauthy/releases/tag/v0.27.0
     DB::hql().shutdown().await.unwrap();
 
     Ok(())
-}
-
-fn get_http_port() -> String {
-    let port = env::var("LISTEN_PORT_HTTP").unwrap_or_else(|_| "8080".to_string());
-    info!("HTTP listen port: {}", port);
-    port
-}
-
-fn get_https_port() -> String {
-    let port = env::var("LISTEN_PORT_HTTPS").unwrap_or_else(|_| "8443".to_string());
-    info!("HTTPS listen port: {}", port);
-    port
-}
-
-fn default_headers() -> middleware::DefaultHeaders {
-    middleware::DefaultHeaders::new()
-        .add(("x-frame-options", "DENY"))
-        .add(("x-content-type-options", "nosniff"))
-        .add((
-            "strict-transport-security",
-            "max-age=31536000;includeSubDomains",
-        ))
-        .add(("referrer-policy", "no-referrer"))
-        .add(("x-robots-tag", "noindex, nofollow"))
-        .add((
-            "content-security-policy",
-            "frame-ancestors 'none'; object-src 'none';",
-        ))
-        .add(("cache-control", "no-store"))
-}
-
-fn api_services() -> actix_web::Scope {
-    web::scope("/auth")
-        .service(generic::redirect_v1)
-        .service(users::get_user_webid)
-        .service(
-            web::scope("/v1")
-                .service(api_keys::get_api_keys)
-                .service(api_keys::post_api_key)
-                .service(api_keys::put_api_key)
-                .service(api_keys::delete_api_key)
-                .service(api_keys::get_api_key_test)
-                .service(api_keys::put_api_key_secret)
-                .service(auth_providers::post_providers)
-                .service(auth_providers::get_providers_minimal)
-                .service(auth_providers::post_provider)
-                .service(auth_providers::post_provider_login)
-                .service(auth_providers::get_provider_delete_safe)
-                .service(auth_providers::post_provider_lookup)
-                .service(auth_providers::get_provider_callback_html)
-                .service(auth_providers::post_provider_callback)
-                .service(auth_providers::delete_provider_link)
-                .service(auth_providers::put_provider)
-                .service(auth_providers::delete_provider)
-                .service(auth_providers::get_provider_img)
-                .service(auth_providers::put_provider_img)
-                .service(auth_providers::post_provider_link)
-                .service(blacklist::get_blacklist)
-                .service(blacklist::post_blacklist)
-                .service(blacklist::delete_blacklist)
-                .service(events::post_events)
-                .service(events::sse_events)
-                .service(events::post_event_test)
-                .service(dev_only::get_template)
-                .service(dev_only::post_dev_only_endpoints)
-                .service(html::get_index)
-                .service(html::get_account_html)
-                .service(html::get_admin_html)
-                .service(html::get_admin_api_keys_html)
-                .service(html::get_admin_attr_html)
-                .service(html::get_admin_blacklist_html)
-                .service(html::get_admin_clients_html)
-                .service(html::get_admin_config_argon2_html)
-                .service(html::get_admin_config_encryption_html)
-                .service(html::get_admin_config_jwks_html)
-                .service(html::get_admin_config_policy_html)
-                .service(html::get_admin_docs_html)
-                .service(html::get_admin_events_html)
-                .service(html::get_admin_groups_html)
-                .service(html::get_admin_roles_html)
-                .service(html::get_admin_scopes_html)
-                .service(html::get_admin_providers_html)
-                .service(html::get_admin_sessions_html)
-                .service(html::get_admin_users_html)
-                .service(html::get_device_html)
-                .service(html::get_fed_cm_html)
-                .service(generic::get_auth_check)
-                .service(generic::get_auth_check_admin)
-                .service(generic::post_update_language)
-                .service(generic::get_version)
-                .service(generic::get_whoami)
-                .service(oidc::get_authorize)
-                .service(oidc::post_authorize)
-                .service(oidc::post_authorize_refresh)
-                .service(oidc::post_device_auth)
-                .service(oidc::post_device_verify)
-                .service(oidc::get_callback_html)
-                .service(oidc::get_certs)
-                .service(oidc::get_cert_by_kid)
-                .service(oidc::get_logout)
-                .service(oidc::post_logout)
-                .service(oidc::rotate_jwk)
-                .service(oidc::post_session)
-                .service(oidc::get_session_info)
-                .service(oidc::get_session_xsrf)
-                .service(clients::get_clients)
-                .service(clients::get_client_by_id)
-                .service(clients::get_client_logo)
-                .service(clients::put_client_logo)
-                .service(clients::delete_client_logo)
-                .service(clients::get_client_secret)
-                .service(clients::post_clients)
-                .service(clients::put_clients)
-                .service(clients::put_generate_client_secret)
-                .service(clients::delete_client)
-                .service(clients::post_clients_dyn)
-                .service(clients::get_clients_dyn)
-                .service(clients::put_clients_dyn)
-                .service(generic::get_login_time)
-                .service(fed_cm::get_fed_cm_accounts)
-                .service(fed_cm::get_fed_cm_config)
-                .service(fed_cm::get_fed_cm_client_meta)
-                .service(fed_cm::get_fed_cm_well_known)
-                .service(fed_cm::post_fed_cm_token)
-                .service(fed_cm::get_fed_client_config)
-                .service(fed_cm::get_fed_cm_status)
-                .service(users::get_users)
-                .service(users::get_users_register)
-                .service(users::post_users_register)
-                .service(users::get_cust_attr)
-                .service(users::post_cust_attr)
-                .service(users::put_cust_attr)
-                .service(users::delete_cust_attr)
-                .service(users::get_user_picture_config)
-                .service(users::get_user_by_id)
-                .service(users::get_user_attr)
-                .service(users::put_user_attr)
-                .service(users::put_user_picture)
-                .service(users::get_user_picture)
-                .service(users::delete_user_picture)
-                .service(users::get_user_devices)
-                .service(users::put_user_device_name)
-                .service(users::delete_user_device)
-                .service(users::get_user_webid_data)
-                .service(users::put_user_webid_data)
-                .service(users::get_user_email_confirm)
-                .service(users::post_user_self_convert_passkey)
-                .service(generic::post_password_hash_times)
-                .service(sessions::get_sessions)
-                .service(sessions::delete_sessions)
-                .service(sessions::delete_session_by_id)
-                .service(sessions::delete_sessions_for_user)
-                .service(users::get_user_password_reset)
-                .service(users::put_user_password_reset)
-                .service(users::get_user_by_email)
-                .service(users::post_users)
-                .service(users::put_user_by_id)
-                .service(users::put_user_self)
-                .service(users::delete_user_by_id)
-                .service(users::post_user_password_request_reset)
-                .service(users::get_user_webauthn_passkeys)
-                .service(users::post_webauthn_reg_start)
-                .service(users::post_webauthn_reg_finish)
-                .service(users::post_webauthn_auth_start)
-                .service(users::post_webauthn_auth_finish)
-                .service(users::delete_webauthn)
-                .service(generic::get_password_policy)
-                .service(generic::put_password_policy)
-                .service(generic::post_pow)
-                .service(generic::get_search)
-                .service(groups::get_groups)
-                .service(groups::post_group)
-                .service(groups::put_group)
-                .service(groups::delete_group)
-                .service(roles::get_roles)
-                .service(roles::post_role)
-                .service(roles::put_role)
-                .service(roles::delete_role)
-                .service(scopes::get_scopes)
-                .service(scopes::post_scope)
-                .service(scopes::put_scope)
-                .service(scopes::delete_scope)
-                .service(oidc::post_token)
-                .service(oidc::post_token_introspect)
-                .service(oidc::get_userinfo)
-                .service(oidc::post_userinfo)
-                .service(oidc::get_forward_auth)
-                .service(generic::get_enc_keys)
-                .service(generic::post_migrate_enc_key)
-                .service(generic::ping)
-                .service(themes::get_theme)
-                .service(themes::post_theme)
-                .service(themes::put_theme)
-                .service(themes::delete_theme)
-                .service(oidc::post_validate_token)
-                .service(oidc::get_well_known)
-                .service(generic::get_health)
-                .service(generic::get_i18n_config)
-                .service(generic::get_ready)
-                .service(html::get_static_assets),
-        )
 }
